Support describing of a partitioned table
